update source code

+++ src/main/java/com/example/springsecurityjwt/service/ItemService.java
@@ -3,5 +3,5 @@ package com.example.springsecurityjwt.service;
 import com.example.springsecurityjwt.model.entities.Item;
 import com.example.springsecurityjwt.service.base.RestService;
 
-public interface ItemService extends RestService<Item, Integer> {
+public interface ItemService extends RestService<Item, Long> {
 }
